refactor: move DashCardTabMenu to DashCardActionsPanel
