+++ frontend/src/metabase/dashboard/components/DashCard/DashCardActionsPanel/DashCardTabMenu/index.ts
